pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qqr"
version = "0.1.0"
authors = [{ name = "boli.cbl", email = "boli.cbl@alibaba-inc.com" }]
description = "A lightweight and efficient RL framework."
readme = { file = "README.md", content-type = "text/markdown" }
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = ["openai-agents>=0.6.0"]

[project.urls]
Homepage = "https://alibaba-nlp.github.io/qqr/"
Repository = "https://github.com/Alibaba-NLP/qqr"
